fix(data): eliminate JAX dtype mismatch warning in diagonal correction
Fixed float64→float32 casting warning in _correct_diagonal method by
explicitly matching dtype of intermediate arrays to c2_mat.dtype.

Problem:
- jnp.zeros(size) and jnp.ones(size) default to float64
- c2_mat may be float32 (common in GPU operations)
- Scatter operation at line 1023 raised FutureWarning:
  "scatter inputs have incompatible types: cannot safely cast value
   from dtype=float64 to dtype=float32"

Solution:
- Line 1016: jnp.zeros(size, dtype=c2_mat.dtype)
- Line 1019: jnp.ones(size, dtype=c2_mat.dtype)
- Ensures diag_val and norm match c2_mat dtype

Verification:
- All 4 verification tests passed
- OLD code: produces expected dtype warning
- FIXED code: eliminates dtype warning with float32/float64
- Output dtype correctly matches input dtype

Impact:
- Eliminates JAX dtype casting warning
- Prevents potential future JAX errors
- Maintains numerical correctness across float32/float64

+++ homodyne/data/xpcs_loader.py
@@ -1013,10 +1013,10 @@ class XPCSDataLoader:
 
         # Create diagonal values using the same array type as input
         if HAS_JAX and hasattr(c2_mat, "device"):  # JAX array
-            diag_val = jnp.zeros(size)
+            diag_val = jnp.zeros(size, dtype=c2_mat.dtype)
             diag_val = diag_val.at[:-1].add(side_band)
             diag_val = diag_val.at[1:].add(side_band)
-            norm = jnp.ones(size)
+            norm = jnp.ones(size, dtype=c2_mat.dtype)
             norm = norm.at[1:-1].set(2)
             # Update diagonal using JAX immutable operations
             diag_indices = np.diag_indices(size)
